Use size_t instead of int to denote array size in maximum subarray

+++ dynamic_programming/maximum_subarray/subarray.cpp
@@ -2,7 +2,7 @@
 
 #include "subarray.h"
 
-int max_subarray(int *a, int n) {
+int max_subarray(int *a, size_t n) {
     if (!n) {
         return 0;
     }
